requirements-dev.txt
PyOpenGL>=3.1.7
pytest>=7.0
